// src/queries/__generated__/USER_ADD_COUPON_MUTATION.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL mutation operation: USER_ADD_COUPON_MUTATION
// ====================================================

export interface USER_ADD_COUPON_MUTATION_userAddCoupon {
  __typename: "SellOrder";
  id: string;
}

export interface USER_ADD_COUPON_MUTATION {
  userAddCoupon: USER_ADD_COUPON_MUTATION_userAddCoupon | null;
}

export interface USER_ADD_COUPON_MUTATIONVariables {
  code: string;
}